Refactor sim arch
