+++ src/sim/strategies/strategy.py
@@ -1,20 +1,21 @@
 """
-Provides the baseline scenario strategy.
+Provides the base scenario template.
 """
+from abc import ABC, abstractmethod
 from typing import List, Type
-from ...scenario import Scenario
-from ...processing import SingleSimProcessor
-from ...results import SingleSimResults
-from ....metrics import Metric
-from ....logging import get_logger
+from ..scenario import Scenario
+from ..processing import SingleSimProcessor
+from ..results import SingleSimResults
+from ...metrics import Metric
+from ...logging import get_logger
 
 logger = get_logger(__name__)
 
 
 # pylint: disable=too-few-public-methods
-class BaselineStrategy:
+class Strategy(ABC):
     """
-    Strategy for simulating the baseline scenario.
+    Base strategy implements the __call__ method.
     """
 
     def __init__(self, metrics: List[Type[Metric]]):
@@ -27,8 +28,7 @@ class BaselineStrategy:
         i: int | None = None,
     ) -> SingleSimResults:
         """
-        Takes in a mutable scenario object and executes
-        the baseline risk strategy on it.
+        Runs the scenario.
         """
         logger.info("STARTING new simulation %d", i)
 
@@ -41,9 +41,12 @@ class BaselineStrategy:
         for sample in scenario.pricepaths:
             scenario.prepare_for_trades(sample)
             scenario.perform_actions(sample)
-            # scenario.after_trades()
             processor.update(sample.timestamp, inplace=True)
 
         logger.info("DONE with simulation %d", i)
 
         return processor.process()
+
+    @abstractmethod
+    def apply_shocks(self, scenario_template: Scenario) -> None:
+        """Apply the scenario shocks to the template."""
